pyproject.toml
[project]
name = "news-summaries"
version = "0.1.0"
description = "AI-powered news summarization service"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.50.0",
    "feedparser>=6.0.0",
    "newspaper3k>=0.2.8",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
